[style] slice 파일 구조 변경
- slice, thunk, types 파일 세개로 분리
