+++ frontend/src/redux/modules/authSlice/types.ts
@@ -1,0 +1,7 @@
+export interface AuthState {
+  user: UserInfo | null;
+  loading: boolean;
+  error: boolean;
+  isLogin: boolean;
+  role: string;
+}
