+++ frontend/src/redux/modules/userSlice/types.ts
@@ -1,0 +1,6 @@
+export interface UserState {
+  userAddress: UserAddressInfo;
+  loading: boolean;
+  error: boolean;
+  isUpdate: boolean;
+}
